requirements.txt
streamlit
pdfplumber
python-docx
spacy
numpy
PyPDF2
scikit-learn
fuzzywuzzy
rapidfuzz
pandas
python-Levenshtein
nltk
phonenumbers